+++ docs/CHANGES.rst
@@ -7,6 +7,8 @@ v0.4.1, ????-??-??
 
 * `Added support for code directive
   <https://github.com/gradha/quicklook-rest-with-nimrod/issues/33>`_.
+* `Uses local time in preview HTML
+  <https://github.com/gradha/quicklook-rest-with-nimrod/issues/34>`_.
 
 v0.4.0, 2014-02-28
 ------------------